apps/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.0
web3>=6.0
celery>=5.3
redis>=5.0
Pillow>=10.0
djangorestframework-simplejwt
google-auth
requests
